+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
  * Bump spring-boot-dependencies from 3.3.0 to 3.3.1
 
+### Features
+
+ * Add questionnaire system
+
 ## 1.6.6
 
 ### Bug Fixes